Network & Disk performance extended

+++ Hardware/CPU/AMD0FCPU.cs
@@ -142,16 +142,16 @@ namespace OpenHardwareMonitor.Hardware.CPU {
             double curMP = 0.5 * ((eax & 0x3F) + 8);
             double maxMP = 0.5 * ((eax >> 16 & 0x3F) + 8);
             coreClocks[i].Value = 
-              (float)(curMP * TimeStampCounterFrequency / maxMP);
-            newBusClock = (float)(TimeStampCounterFrequency / maxMP);
+              (curMP * TimeStampCounterFrequency / maxMP);
+            newBusClock = (TimeStampCounterFrequency / maxMP);
           } else {
             // Fail-safe value - if the code above fails, we'll use this instead
-            coreClocks[i].Value = (float)TimeStampCounterFrequency;
+            coreClocks[i].Value = TimeStampCounterFrequency;
           }
         }
 
         if (newBusClock > 0) {
-          this.busClock.Value = (float)newBusClock;
+          this.busClock.Value = newBusClock;
           ActivateSensor(this.busClock);
         }
       }
